+++ projects/ui.voila.dev/apps/ui.voila.dev/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ui.voila.dev/ui.voila.dev
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @voila.dev/ui@1.0.5
+
 ## 0.0.1
 
 ### Patch Changes